refactor(domain): move props to entity files and remove ports directories
