+++ src/application/usecases/answer-question/answer-question.usecase.ts
@@ -2,8 +2,7 @@ import type { UseCase } from '@/core/application/use-case'
 import type { AnswersRepository } from '@/application/repositories/answers.repository'
 import type { QuestionsRepository } from '@/application/repositories/questions.repository'
 import type { UsersRepository } from '@/application/repositories/users.repository'
-import { Answer } from '@/domain/entities/answer/answer.entity'
-import type { AnswerProps } from '@/domain/entities/answer/ports/answer.props'
+import { Answer, type AnswerProps } from '@/domain/entities/answer/answer.entity'
 import { ResourceNotFoundError } from '@/shared/application/errors/resource-not-found.error'
 import type { OmitTimestamps } from '@/shared/types/custom/omit-timestamps'
