+++ src/application/usecases/create-account/create-account.usecase.ts
@@ -1,8 +1,7 @@
 import type { UseCase } from '@/core/application/use-case'
 import type { UsersRepository } from '@/application/repositories/users.repository'
 import type { PasswordHasher } from '@/infra/adapters/security/ports/password-hasher'
-import type { UserProps } from '@/domain/entities/user/ports/user.props'
-import { User } from '@/domain/entities/user/user.entity'
+import { User, type UserProps } from '@/domain/entities/user/user.entity'
 import { UserWithEmailAlreadyRegisteredError } from './errors/user-with-email-already-registered.error'
 import type { OmitTimestamps } from '@/shared/types/custom/omit-timestamps'
