+++ src/application/usecases/create-question/create-question.usecase.ts
@@ -1,7 +1,6 @@
 import type { UseCase } from '@/core/application/use-case'
 import type { QuestionsRepository } from '@/application/repositories/questions.repository'
-import type { QuestionProps } from '@/domain/entities/question/ports/question.props'
-import { Question } from '@/domain/entities/question/question.entity'
+import { Question, type QuestionProps } from '@/domain/entities/question/question.entity'
 import { QuestionWithTitleAlreadyRegisteredError } from './errors/question-with-title-already-registered.error'
 import type { OmitTimestamps } from '@/shared/types/custom/omit-timestamps'
