--- src/domain/entities/answer-comment/ports/answer-comment.props.ts
@@ -1,5 +1,0 @@
-import type { CommentProps } from '@/domain/entities/comment/ports/comment.props'
-
-export type AnswerCommentProps = CommentProps & {
-  answerId: string
-}
