--- src/domain/entities/comment/ports/comment.props.ts
@@ -1,4 +1,0 @@
-import type { Comment } from '../comment.entity'
-import type { Props } from '@/shared/types/custom/props'
-
-export type CommentProps = Props<typeof Comment>
