--- src/domain/entities/email-validation/ports/email-validation.props.ts
@@ -1,8 +1,0 @@
-import type { EmailValidationCode } from '@/domain/value-objects/email-validation-code/email-validation-code.vo'
-
-export type EmailValidationProps = {
-  email: string
-  code: EmailValidationCode
-  expiresAt: Date
-  isVerified: boolean
-}
