--- src/domain/entities/question-comment/ports/question-comment.props.ts
@@ -1,5 +1,0 @@
-import type { CommentProps } from '@/domain/entities/comment/ports/comment.props'
-
-export type QuestionCommentProps = CommentProps & {
-  questionId: string
-}
