+++ src/domain/entities/question-comment/question-comment.entity.ts
@@ -1,5 +1,8 @@
-import { Comment } from '../comment/comment.entity'
-import type { QuestionCommentProps } from './ports/question-comment.props'
+import { Comment, type CommentProps } from '../comment/comment.entity'
+
+export type QuestionCommentProps = CommentProps & {
+  questionId: string
+}
 
 export class QuestionComment extends Comment {
   readonly questionId: string
