--- src/domain/entities/question/ports/question.props.ts
@@ -1,5 +1,0 @@
-import type { Question } from '../question.entity'
-import type { Optional } from '@/shared/types/common/optional'
-import type { Props } from '@/shared/types/custom/props'
-
-export type QuestionProps = Optional<Omit<Props<typeof Question>, 'slug' | 'answers'>, 'bestAnswerId'>
