--- src/domain/entities/user/ports/user.props.ts
@@ -1,4 +1,0 @@
-import type { User } from '../user.entity'
-import type { Props } from '@/shared/types/custom/props'
-
-export type UserProps = Props<typeof User>
